# api/swagger.yaml
openapi: 3.0.0
info:
  title: Infinity Servers API
  description: This is a simple API for K8
  contact:
    email: support@infinityservers.co.uk
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/Infinityservers/GoK8/1.0.0
  description: SwaggerHub API Auto Mocking
tags:
- name: namespace
  description: Namespace based calls.
- name: deployment
  description: Deployment based calls.
paths:
  /namespace:
    get:
      tags:
      - namespace
      summary: Get a list of namespaces
      description: |
        By passing in the appropriate options, you can create a new namespace with tags
      operationId: ListNamespaces
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Namespace'
                x-content-type: application/json
        "400":
          description: bad input parameter
  /namespace/{name}:
    get:
      tags:
      - namespace
      summary: Get a single namespace
      description: |
        By passing in the appropriate options, you can create a new namespace with tags
      operationId: GetNamespace
      parameters:
      - name: name
        in: path
        description: Get a single namespace
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Namespace'
                x-content-type: application/json
        "400":
          description: bad input parameter
    post:
      tags:
      - namespace
      summary: Create a new namespace
      description: |
        By passing in the appropriate options, you can create a new namespace with tags
      operationId: CreateNamespace
      parameters:
      - name: name
        in: path
        description: name of the namespace you wish to create
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Namespace'
                x-content-type: application/json
        "400":
          description: bad input parameter
    delete:
      tags:
      - namespace
      summary: Delete an existing namespace
      description: |
        By passing in the appropriate options, you can create a new namespace with tags
      operationId: DeleteNamespace
      parameters:
      - name: name
        in: path
        description: name of the namespace you wish to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Namespace'
                x-content-type: application/json
        "400":
          description: bad input parameter
  /deployment:
    get:
      tags:
      - deployment
      summary: Get a list of deployments in namespace
      description: |
        By passing in the appropriate options, you can create a new namespace with tags
      operationId: ListDeployments
      parameters:
      - name: namespace
        in: query
        description: Namespace you want a deployment list for
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
                x-content-type: application/json
        "400":
          description: bad input parameter
  /deployment/{deployment}:
    get:
      tags:
      - deployment
      summary: Get a single deployment
      description: |
        By passing in the appropriate options, you can create a new namespace with tags
      operationId: GetDeployment
      parameters:
      - name: deployment
        in: path
        description: Deployment you wish to get info on
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: namespace
        in: query
        description: Namespace the deployment is in
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
                x-content-type: application/json
        "400":
          description: bad input parameter
    post:
      tags:
      - deployment
      summary: Create a new deployment
      description: |
        By passing in the appropriate options, you can create a new namespace with tags
      operationId: CreateDeployment
      parameters:
      - name: deployment
        in: path
        description: Deployment name you wish to set
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: namespace
        in: query
        description: Namespace the deployment is in
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: app
        in: query
        description: App name you wish to set
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: image
        in: query
        description: The image you wish to deploy
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
                x-content-type: application/json
        "400":
          description: bad input parameter
    delete:
      tags:
      - deployment
      summary: Delete a single deployment
      description: |
        By passing in the appropriate options, you can create a new namespace with tags
      operationId: DeleteDeployment
      parameters:
      - name: deployment
        in: path
        description: Deployment you wish to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: namespace
        in: query
        description: Namespace the deployment is in
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
                x-content-type: application/json
        "400":
          description: bad input parameter
components:
  schemas:
    Namespace:
      type: object
      properties:
        name:
          type: string
          example: test-name
      example:
        name: test-name
    Deployment:
      type: object
      properties:
        deployment-name:
          type: string
          example: nginx-deployment
        app:
          type: string
          example: nginx
        image:
          type: string
          example: nginx:1.14.2
      example:
        app: nginx
        image: nginx:1.14.2
        deployment-name: nginx-deployment