# go/model_namespace.go
/*
 * Infinity Servers API
 *
 * This is a simple API for K8
 *
 * API version: 1.0.0
 * Contact: support@infinityservers.co.uk
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Namespace struct {

	Name string `json:"name,omitempty"`
}